+++ lib/kong_schema/functional.rb
@@ -3,5 +3,13 @@ module KongSchema
     def blank?(object)
       object.nil? || object.empty?
     end
+
+    def try(recv, meth)
+      if recv.nil?
+        nil
+      else
+        recv.send(meth)
+      end
+    end
   end
 end