+++ lib/kong_schema/resource.rb
@@ -1,3 +1,4 @@
 require_relative './resource/api'
+require_relative './resource/plugin'
 require_relative './resource/target'
 require_relative './resource/upstream'