+++ spec/kong_schema/resource/upstream_spec.rb
@@ -2,10 +2,6 @@ describe KongSchema::Resource::Upstream do
   let(:schema) { KongSchema::Schema }
   let(:test_utils) { KongSchemaTestUtils.new }
 
-  after(:each) do
-    test_utils.reset_kong
-  end
-
   describe 'creating upstreams' do
     let :config do
       test_utils.generate_config({