+++ spec/support/kong_schema_test_utils.rb
@@ -1,4 +1,5 @@
 require 'kong_schema'
+require 'fileutils'
 
 class KongSchemaTestUtils
   attr_reader :host
@@ -9,6 +10,38 @@ class KongSchemaTestUtils
 
   def generate_config(config = {})
     JSON.parse(JSON.dump({ admin_host: host }.merge(config)))
+  end
+
+  def generate_config_file(config = {}, format: :yaml)
+    buffer = case format
+    when :json
+      JSON.dump(config)
+    else
+      YAML.dump(config)
+    end
+
+    filename = case format
+    when :json
+      'config.json'
+    else
+      'config.yaml'
+    end
+
+    filepath = File.join(Dir.pwd, 'spec', 'fixtures', filename)
+
+    File.write(filepath, buffer)
+    yield filepath
+  ensure
+    FileUtils.rm(filepath)
+  end
+
+  def fake_stdin(*args)
+    $stdin = StringIO.new
+    $stdin.puts(args.shift) until args.empty?
+    $stdin.rewind
+    yield
+  ensure
+    $stdin = STDIN
   end
 
   def reset_kong